Add WSL build target and fix three compile/correctness errors
do_compile.sh: new `wsl` target for Ubuntu-on-WSL with Intel oneAPI 2026
(Intel MPI mpicxx with a g++ backend, MKL rpath, I_MPI_FABRICS=shm). The
existing idm3 / idm4 / nurion / mac targets are untouched.

Three source fixes were needed. The first two are only exposed by this
target because no previous one uses a GCC backend:

  src/bath.cpp:646,670
      (double[]){0.0,0.0,0.0} is a C99 compound literal. GCC's C++
      frontend treats it as a prvalue temporary, so the array-to-pointer
      decay fails with "error: taking address of temporary array".
      mpiicpc and clang accept it as an extension, which is why this
      never surfaced before. Replaced with a named local array;
      BathArray_setBath_i_xyz copies element-wise, so behavior and
      numerics are unchanged.

  src/simulator_gcce.cpp:267
      pulse->total_pulse_axes is never declared anywhere in the tree.
      The actual member is Pulse::pulse_axes (include/pulse.h:40).
      This is a hard error on every compiler.

  src/qubit.cpp:246
      qgyro was declared float while QubitArray_getQubit_i_gyro returns
      double, truncating the qubit gyromagnetic ratio before it reaches
      calZeemanVector. For the NV this is -17608.59705 -> -17608.59765625
      (6.06e-04 radkHz/G). Unlike the two above, this one changes results:
      measured max|dL| is 0 for CCE, which never evaluates the qubit
      Zeeman term, and up to 7.97e-04 for gCCE, which does.

Also adds a .gitignore, which the repo did not have. It covers build
artifacts only. Two deliberate omissions: *.out is NOT ignored, because
60+ DFT/QE outputs under example/ are tracked; and *.process is exempted
under test/regression/, because those are reference outputs.

+++ src/qubit.cpp
@@ -243,7 +243,7 @@ MatrixXcd QubitArray_ZeemanHamil(QubitArray* qa, MatrixXcd** sigmas, int iq, flo
         exit(EXIT_FAILURE);
     }
     
-    float qgyro = QubitArray_getQubit_i_gyro(qa,iq);
+    double qgyro = QubitArray_getQubit_i_gyro(qa,iq);
     MatrixXcd vecZeeman = calZeemanVector(qgyro,bfield);
 
     return calHamiltonianSingleInt(vecZeeman,sigmas[iq]);
